Resoning Agent/requirements.txt
# ResQNet Reasoning Agent — Dependencies
twikit
httpx
fastapi
uvicorn[standard]
apscheduler
pydantic
pydantic-settings
python-dotenv
loguru
